Made elementary stocks, fixed early corporation
Lots of changes. Made an elementary stock script that works without the 4s data, and also sends messages for other scripts to see which companies should be worked for. Also made a stockbuying script that works with or without the 4s data.

Also fixed a whole bunch of changes to the corporation script to make it work better in the early stage.